# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Sales AI System API
  version: 1.0.0
  description: Browser clients authenticate with an opaque server-side session cookie. Unsafe authenticated requests also require the X-CSRF-Token returned by GET /api/auth/me.
servers:
  - url: http://localhost:3000
security:
  - cookieSession: []
paths:
...
  /api/leads:
    get:
      summary: List sales leads
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/LeadLimit'
        - name: keyword
          in: query
          description: Case-insensitive match against company name, project title, URL, description, lead reason, or owner memo.
          schema: { type: string, maxLength: 200 }
        - name: source
          in: query
          description: Project platform type. It does not use the internal SalesLead source label.
          schema: { $ref: '#/components/schemas/PlatformType' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/LeadStatus' }
        - name: priority
          in: query
          schema: { $ref: '#/components/schemas/LeadPriority' }
        - name: contactState
          in: query
          description: "`has` requires a legacy lead contact value or an active company contact with email or inquiry URL. Deleted and unsubscribed company contacts are ignored; `none` has neither."
          schema: { type: string, enum: [has, none] }
        - name: mailStatus
          in: query
          description: Status of the latest mail for the lead, ordered by createdAt then id descending. `none` selects leads without any mail.
          schema: { type: string, enum: [none, draft, in_review, rejected, approved, queued, sending, sent, failed, cancelled] }
        - name: nextAction
          in: query
          description: "`any` applies no next-action filter. `scheduled` has a next-action date, follow-up date, or active task. `overdue` has one due at or before the server time. `none` has none of those values."
          schema: { type: string, enum: [any, scheduled, overdue, none] }
        - name: sort
          in: query
          schema: { type: string, enum: [company, project, amount, supporters, daysLeft, score, priority, createdAt], default: createdAt }
        - name: sortDirection
          in: query
          description: Defaults to `desc` for `createdAt` and `asc` for other sort fields.
          schema: { type: string, enum: [asc, desc] }
      responses:
        '200':
          description: Lead list. `data.summary` is calculated after source, status, priority, and nextAction filters, before contactState and mailStatus summary selections.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/LeadListData' }
    post:
      summary: Create a sales lead
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateLeadRequest' }
      responses:
        '201':
          description: Lead response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /api/leads/today:
    get:
      summary: List today's actionable sales leads
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated today and overdue work with category counts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/leads/{id}:
    get:
      summary: Get a sales lead
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Lead response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    patch:
      summary: Update a sales lead
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateLeadRequest' }
      responses:
        '200':
          description: Lead response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/leads/{id}/score:
    post:
      summary: Recalculate lead score
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: Score response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /api/ai/leads/{leadId}/generate-mail:
    post:
      summary: Generate a mail draft for a lead
      parameters:
        - $ref: '#/components/parameters/LeadId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateMailRequest' }
      responses:
        '201':
          description: AI response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '409':
          description: Analysis is unconfirmed, stale, belongs to another project, or is not the latest confirmed version
  /api/ai/leads/{leadId}/email-draft:
    post:
      summary: Generate a mail draft alias for a lead
      parameters:
        - $ref: '#/components/parameters/LeadId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateMailRequest' }
      responses:
        '201':
          description: AI response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '409':
          description: Analysis is unconfirmed, stale, belongs to another project, or is not the latest confirmed version
  /api/ai/leads/{leadId}/analysis:
    get:
      summary: Get structured lead analysis and revision history
      parameters:
        - $ref: '#/components/parameters/LeadId'
      responses:
        '200':
          description: Structured lead analysis response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LeadAnalysisResponse'
    patch:
      summary: Save a structured lead analysis draft revision
      parameters:
        - $ref: '#/components/parameters/LeadId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeadAnalysisRevisionRequest' }
      responses:
        '200':
          description: Saved structured lead analysis draft
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LeadAnalysisResponse'
        '409':
          description: Version conflict or the lead/project changed while editing
  /api/ai/leads/{leadId}/analysis/confirm:
    post:
      summary: Confirm a structured lead analysis revision
      parameters:
        - $ref: '#/components/parameters/LeadId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeadAnalysisRevisionRequest' }
      responses:
        '201':
          description: Confirmed structured lead analysis revision
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LeadAnalysisResponse'
        '409':
          description: Version conflict, stale project data, or one or more required analysis values are missing
  /api/ai/leads/{leadId}/analyze:
    post:
      summary: Analyze a lead
      parameters:
        - $ref: '#/components/parameters/LeadId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: AI response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/ai/mails/{mailId}/polish:
    post:
      summary: Polish a mail
      parameters:
        - $ref: '#/components/parameters/MailId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: AI response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '429':
          description: OpenAI monthly budget would be exceeded before this request
  /api/ai/mails/{mailId}/semantic-consistency:
    post:
      summary: Check mail semantic consistency
      parameters:
        - $ref: '#/components/parameters/MailId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: AI response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '429':
          description: OpenAI monthly budget would be exceeded before this request
  /api/ai/leads/{leadId}/generations:
    get:
      summary: List AI generations for a lead
      parameters:
        - $ref: '#/components/parameters/LeadId'
      responses:
        '200':
          description: AI generation response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /api/ai/replies/{replyId}/classify:
    post:
      summary: Classify a reply
      parameters:
        - $ref: '#/components/parameters/ReplyId'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '201':
          description: AI response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: sales_ai_session
      description: Opaque server-managed session. Production uses the __Host-sales_ai_session cookie name.
...
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, default: 20 }
    LeadLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    LeadId:
      name: leadId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    MailId:
      name: mailId
      in: path
      required: true
      schema: { type: string }
...
    ReplyId:
      name: replyId
      in: path
      required: true
      schema: { type: string }
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Required on every unsafe authenticated request. Obtain it from GET /api/auth/me and keep it in memory only.
      schema: { type: string }
  schemas:
    ApiResponse:
      type: object
      required: [data, meta, error]
      properties:
        data: {}
        meta:
          type: object
          nullable: true
          additionalProperties: true
        error:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ErrorResponse'
...
    LeadListData:
      type: object
      required: [items, page, limit, total, summary]
      properties:
        items:
          type: array
          description: Each lead retains its existing fields, includes its newest mail with only id, leadId, companyId, status, and createdAt, and includes active `company.contacts` with only id, email, inquiryUrl, isPrimary, isUnsubscribed, and deletedAt.
          items:
            type: object
            additionalProperties: true
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        summary:
          type: object
          required: [total, noContact, draft, review, queued]
          properties:
            total: { type: integer, minimum: 0 }
            noContact: { type: integer, minimum: 0 }
            draft: { type: integer, minimum: 0 }
            review: { type: integer, minimum: 0 }
            queued: { type: integer, minimum: 0 }
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    PlatformType:
      type: string
      enum: [campfire, makuake, green_funding, other]
    ProjectStatus:
      type: string
      enum: [discovered, active, ended, suspended, unknown]
    LeadStatus:
      type: string
      enum: [discovered, qualified, drafted, reviewing, approved, queued, contacted, replied, meeting_candidate, rejected, no_response, archived]
    LeadPriority:
      type: string
      enum: [high, medium, low]
...
    CreateLeadRequest:
      type: object
      required: [companyId]
      properties:
        companyId: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        source: { type: string }
        ownerMemo: { type: string }
        nextActionAt: { type: string, format: date-time }
        contactEmail: { type: string }
        contactFormUrl: { type: string }
        siteMessageUrl: { type: string }
        contactMemo: { type: string }
        sendMethod: { type: string }
        sentAt: { type: string, format: date-time }
        nextFollowUpAt: { type: string, format: date-time }
        brandWebsiteUrl: { type: string }
        instagramUrl: { type: string }
        tiktokUrl: { type: string }
        xUrl: { type: string }
        brandAnalysisMemo: { type: string }
        snsAnalysisMemo: { type: string }
    UpdateLeadRequest:
      type: object
      properties:
        status: { $ref: '#/components/schemas/LeadStatus' }
        priority: { $ref: '#/components/schemas/LeadPriority' }
        ownerMemo: { type: string, nullable: true }
        nextActionAt: { type: string, format: date-time, nullable: true }
        contactEmail: { type: string, format: email, nullable: true }
        contactFormUrl: { type: string, format: uri, nullable: true }
        siteMessageUrl: { type: string, format: uri, nullable: true }
        contactMemo: { type: string, nullable: true }
        sendMethod: { type: string, nullable: true }
        sentAt: { type: string, format: date-time, nullable: true }
        nextFollowUpAt: { type: string, format: date-time, nullable: true }
        brandWebsiteUrl: { type: string, format: uri, nullable: true }
        instagramUrl: { type: string, format: uri, nullable: true }
        tiktokUrl: { type: string, format: uri, nullable: true }
        xUrl: { type: string, format: uri, nullable: true }
        brandAnalysisMemo: { type: string, nullable: true }
        snsAnalysisMemo: { type: string, nullable: true }
        leadReason: { type: string, nullable: true }
        companyName: { type: string, minLength: 1 }
        companyWebsiteUrl: { type: string, format: uri, nullable: true }
        companyInquiryUrl: { type: string, format: uri, nullable: true }
        companyIndustry: { type: string, nullable: true }
        companyLocation: { type: string, nullable: true }
        companySourceTotalAmount: { type: integer, minimum: 0, nullable: true }
        companySourceProjectCount: { type: integer, minimum: 0, nullable: true }
        companySourceSupporterCount: { type: integer, minimum: 0, nullable: true }
        companyMemo: { type: string, nullable: true }
        projectTitle: { type: string, minLength: 1 }
        projectUrl: { type: string, format: uri }
        projectSource: { $ref: '#/components/schemas/PlatformType' }
        projectStatus: { $ref: '#/components/schemas/ProjectStatus' }
        projectAmount: { type: integer, minimum: 0 }
        projectSupporterCount: { type: integer, minimum: 0 }
        projectTargetAmount: { type: integer, minimum: 0, nullable: true }
        projectStartDate: { type: string, format: date-time, nullable: true }
        projectEndDate: { type: string, format: date-time, nullable: true }
        projectCategory: { type: string, nullable: true }
        projectLocation: { type: string, nullable: true }
        projectDescription: { type: string, nullable: true }
...
    GenerateMailRequest:
      type: object
      required: [templateKey, analysisRevisionId]
      properties:
        templateKey: { type: string, minLength: 1 }
        tone: { type: string }
        analysisRevisionId: { type: string, format: uuid }
...
    LeadAnalysisRevisionRequest:
      type: object
      required: [expectedVersion, expectedSourceFingerprint, appeal, targetUser, videoIdea]
      properties:
        expectedVersion: { type: integer, minimum: 0 }
        expectedSourceFingerprint: { type: string, minLength: 32, maxLength: 32 }
        appeal: { type: string, maxLength: 1000 }
        targetUser: { type: string, maxLength: 1000 }
        videoIdea: { type: string, maxLength: 1000 }
    LeadAnalysisRevision:
      type: object
      required: [id, leadId, projectId, version, status, origin, sourceFingerprint, humanEdited, editedFields, createdAt]
      properties:
        id: { type: string, format: uuid }
        leadId: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        sourceGenerationId: { type: string, format: uuid, nullable: true }
        changedById: { type: string, format: uuid, nullable: true }
        version: { type: integer, minimum: 1 }
        status:
          type: string
          enum: [draft, confirmed]
        origin:
          type: string
          enum: [generated, manual, migration]
        appeal: { type: string, nullable: true }
        targetUser: { type: string, nullable: true }
        videoIdea: { type: string, nullable: true }
        sourceFingerprint: { type: string, minLength: 1 }
        generatedAt: { type: string, format: date-time, nullable: true }
        confirmedAt: { type: string, format: date-time, nullable: true }
        humanEdited: { type: boolean }
        editedFields:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        stale: { type: boolean }
        missingFields:
          type: array
          items:
            type: string
            enum: [appeal, targetUser, videoIdea]
    LeadAnalysisResponse:
      type: object
      required: [projectId, sourceFingerprint, proposal, confirmed, history, missingFields, stale, canGenerateMail]
      properties:
        projectId: { type: string, format: uuid }
        sourceFingerprint: { type: string, minLength: 32, maxLength: 32 }
        proposal:
          allOf:
            - $ref: '#/components/schemas/LeadAnalysisRevision'
          nullable: true
        confirmed:
          allOf:
            - $ref: '#/components/schemas/LeadAnalysisRevision'
          nullable: true
        history:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/LeadAnalysisRevision'
        missingFields:
          type: array
          items:
            type: string
            enum: [appeal, targetUser, videoIdea]
        stale: { type: boolean }
        canGenerateMail: { type: boolean }
...